+++ cluster-operator/src/test/java/io/strimzi/operator/cluster/operator/assembly/KafkaConnectApiImplTest.java
@@ -7,6 +7,7 @@ package io.strimzi.operator.cluster.operator.assembly;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.github.tomakehurst.wiremock.WireMockServer;
 import com.github.tomakehurst.wiremock.core.WireMockConfiguration;
+import io.strimzi.api.kafka.model.common.ConnectorState;
 import io.strimzi.operator.common.Reconciliation;
 import io.strimzi.test.TestUtils;
 import io.vertx.core.json.JsonObject;
@@ -20,8 +21,14 @@ import java.util.Map;
 import java.util.concurrent.ExecutionException;
 
 import static com.github.tomakehurst.wiremock.client.WireMock.aResponse;
+import static com.github.tomakehurst.wiremock.client.WireMock.equalTo;
 import static com.github.tomakehurst.wiremock.client.WireMock.get;
+import static com.github.tomakehurst.wiremock.client.WireMock.matchingJsonPath;
+import static com.github.tomakehurst.wiremock.client.WireMock.notMatching;
+import static com.github.tomakehurst.wiremock.client.WireMock.post;
+import static com.github.tomakehurst.wiremock.client.WireMock.postRequestedFor;
 import static com.github.tomakehurst.wiremock.client.WireMock.put;
+import static com.github.tomakehurst.wiremock.client.WireMock.urlPathEqualTo;
 import static com.github.tomakehurst.wiremock.client.WireMock.urlPathMatching;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.aMapWithSize;
@@ -174,4 +181,43 @@ public class KafkaConnectApiImplTest {
                         hasEntry("org.apache.kafka.connect", "WARN")))
                 ).toCompletableFuture().join();
     }
+
+    @Test
+    public void testCreateConnectorSendsInitialState() {
+        server.stubFor(post(urlPathEqualTo("/connectors"))
+                .willReturn(aResponse()
+                        .withStatus(201)
+                        .withBody("{\"name\":\"my-connector\",\"config\":{},\"tasks\":[]}")));
+
+        KafkaConnectApi api = new KafkaConnectApiImpl();
+        JsonObject config = new JsonObject().put("connector.class", "Dummy");
+
+        api.createConnector(Reconciliation.DUMMY_RECONCILIATION, "127.0.0.1", server.port(),
+                "my-connector", config, ConnectorState.STOPPED).toCompletableFuture().join();
+
+        // The POST body must carry name, config, and the upper-case initial_state
+        server.verify(postRequestedFor(urlPathEqualTo("/connectors"))
+                .withRequestBody(matchingJsonPath("$.name", equalTo("my-connector")))
+                .withRequestBody(matchingJsonPath("$.initial_state", equalTo("STOPPED")))
+                .withRequestBody(matchingJsonPath("$.config['connector.class']", equalTo("Dummy"))));
+    }
+
+    @Test
+    public void testCreateConnectorWithoutStateOmitsInitialState() {
+        server.stubFor(post(urlPathEqualTo("/connectors"))
+                .willReturn(aResponse()
+                        .withStatus(201)
+                        .withBody("{\"name\":\"my-connector\",\"config\":{},\"tasks\":[]}")));
+
+        KafkaConnectApi api = new KafkaConnectApiImpl();
+
+        api.createConnector(Reconciliation.DUMMY_RECONCILIATION, "127.0.0.1", server.port(),
+                "my-connector", new JsonObject(), null).toCompletableFuture().join();
+
+        // With no requested state, the initial_state field must be omitted (Connect defaults to running)
+        server.verify(postRequestedFor(urlPathEqualTo("/connectors"))
+                .withRequestBody(matchingJsonPath("$.name", equalTo("my-connector")))
+                .withRequestBody(notMatching("(?s).*initial_state.*")));
+    }
+
 }